修复数据

+++ mall-biz/mall-biz-system/src/main/java/com/lingyi/mall/biz/system/service/impl/RoleServiceImpl.java
@@ -4,6 +4,7 @@ import cn.hutool.core.bean.BeanUtil;
 import com.github.pagehelper.PageHelper;
 import com.lingyi.mall.api.system.dto.RoleDTO;
 import com.lingyi.mall.api.system.entity.Role;
+import com.lingyi.mall.api.system.entity.User;
 import com.lingyi.mall.api.system.enums.SystemFailEnum;
 import com.lingyi.mall.api.system.param.RoleParam;
 import com.lingyi.mall.api.system.vo.RoleVO;
@@ -11,13 +12,16 @@ import com.lingyi.mall.biz.system.mapper.RoleMapper;
 import com.lingyi.mall.biz.system.repository.RoleRepository;
 import com.lingyi.mall.biz.system.service.RoleMenuService;
 import com.lingyi.mall.biz.system.service.RoleService;
+import com.lingyi.mall.common.base.exception.BizException;
 import com.lingyi.mall.common.base.param.BasePageParam;
 import com.lingyi.mall.common.base.util.AssertUtil;
+import com.lingyi.mall.common.util.ConverterUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 /**
  * @Author: maweiyan
@@ -29,19 +33,19 @@ import java.util.List;
 @RequiredArgsConstructor
 public class RoleServiceImpl implements RoleService {
 
-    private final RoleRepository mbsRoleRepository;
+    private final RoleRepository roleRepository;
 
-    private final RoleMapper mbsRoleMapper;
+    private final RoleMapper roleMapper;
 
     private final RoleMenuService roleMenuService;
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void add(RoleDTO roleDTO) {
-        //转换
+        //DTO转换Entity
         Role role = BeanUtil.copyProperties(roleDTO, Role.class);
         //保存
-        mbsRoleRepository.save(role);
+        roleRepository.save(role);
         //保存角色菜单信息
         roleMenuService.saveList(role.getId(), roleDTO.getMenuIds());
 
@@ -49,19 +53,26 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public void removeByIds(List<Long> ids) {
-        mbsRoleRepository.deleteAllById(ids);
+        roleRepository.deleteAllById(ids);
     }
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void editById(RoleDTO roleDTO) {
+        //获取id
         Long id = roleDTO.getId();
-        //断言menuId不为空
-        AssertUtil.notNull(id, SystemFailEnum.ROLE_ID_NULL_ERROR);
-        //转换
-        Role role = BeanUtil.copyProperties(roleDTO, Role.class);
+        //获取角色信息
+        Optional<Role> optional = roleRepository.findById(id);
+        //判断用户是否不为空
+        if (optional.isEmpty()) {
+            throw new BizException(SystemFailEnum.ROLE_NULL_ERROR);
+        }
+        //获取用户
+        Role role = optional.get();
+        //DTO转换Entity
+        ConverterUtil.to(roleDTO, role);
         //保存
-        mbsRoleRepository.save(role);
+        roleRepository.save(role);
         //删除角色菜单集
         roleMenuService.removeByRoleId(id);
         //保存角色菜单信息
@@ -70,12 +81,12 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public RoleVO findById(Long id) {
-        return mbsRoleMapper.selectById(id);
+        return roleMapper.selectById(id);
     }
 
     @Override
     public List<RoleVO> findListByPageAndParam(BasePageParam basePageParam, RoleParam roleParam) {
         PageHelper.startPage(basePageParam.getCurrentPage(), basePageParam.getPageSize(), basePageParam.getSort());
-        return mbsRoleMapper.selectListByParam(roleParam);
+        return roleMapper.selectListByParam(roleParam);
     }
 }
